Refactor visualization components to dedicated directory structure
- Created visualizations directory with subcomponents:
  - waves component
  - pill-rating component
  - squares component
  - sparkles-core component
- Added comprehensive documentation with README files and usage examples
- Created consistent directory structure with index.ts exports
- Added backward compatibility re-exports with @deprecated annotations
- Updated COMPONENT-MOVEMENT-PLAN.md to track changes
