+++ src/components/ui/visualizations/rating/index.ts
@@ -1,0 +1,1 @@
+export * from "./pill-rating";
