+++ src/components/ui/visualizations/sparkles/index.ts
@@ -1,0 +1,1 @@
+export * from "./sparkles-core";
