+++ src/components/ui/visualizations/squares/index.ts
@@ -1,0 +1,1 @@
+export * from "./squares";
