+++ src/components/ui/visualizations/waves/index.ts
@@ -1,0 +1,1 @@
+export * from "./waves";
